Shone

+++ TechHealth/View/ManagerView/CRUDRooms/RoomWindow.xaml.cs
@@ -15,33 +15,46 @@ using System.Windows.Shapes;
 using TechHealth.Model;
 using TechHealth.Repository;
 using TechHealth.Controller;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
 
-namespace TechHealth.View.ManagerView
+namespace TechHealth.View.ManagerView.CRUDRooms
 {
-    public partial class ManagerMainWindow : Window
+    public partial class RoomWindow : Window
     {
         private ObservableCollection<Room> rooms;
+        
         private RoomController roomController = new RoomController();
-        public ManagerMainWindow()
+
+        public ObservableCollection<Room> Rooms { get => rooms; set => rooms = value; }
+
+        public RoomWindow()
         {
             InitializeComponent();
+            this.DataContext = this;
             rooms = new ObservableCollection<Room>(RoomRepository.Instance.GetAll().Values);
-            lvUsers.ItemsSource = rooms;
+            //lvUsers.ItemsSource = rooms;
         }
-
+        
         private void Button_Click_Add(object sender, RoutedEventArgs e)
         {
             new AddForm().ShowDialog();
             UpdateView();
-
         }
 
         private void Button_Click_Delete(object sender, RoutedEventArgs e)
         {
-            Room selected = (Room)lvUsers.SelectedItems[0];
-            roomController.Delete(selected.roomId);
-            rooms = new ObservableCollection<Room>(RoomRepository.Instance.GetAll().Values);
-            lvUsers.ItemsSource = rooms;
+            if (lvUsers.SelectedIndex == -1)
+            {
+                MessageBox.Show("You have to select a room before viewing it!");
+            }
+            else
+            {
+                Room selected = (Room)lvUsers.SelectedItems[0];
+                roomController.Delete(selected.roomId);
+                rooms = new ObservableCollection<Room>(RoomRepository.Instance.GetAll().Values);
+                lvUsers.ItemsSource = rooms;
+            }        
         }
 
         private void Button_Click_View(object sender, RoutedEventArgs e)
@@ -50,9 +63,13 @@ namespace TechHealth.View.ManagerView
             {
                 MessageBox.Show("You have to select a room before viewing it!");
             }
-            Room room = (Room)lvUsers.SelectedItems[0];
-            new UpdateForm(room).ShowDialog();
-            UpdateView();
+            else
+            {
+                Room room = (Room)lvUsers.SelectedItems[0];
+                new UpdateForm(room).ShowDialog();
+                UpdateView();
+            }
+                
         }
 
         public void UpdateView()
